已成功修复 GCC 编译器构建产生的所有项目代码警告。
修复内容
类型转换警告 - 添加显式 static_cast 转换
Switch 语句警告 - 补全 ErrorCode 枚举处理
未使用代码警告 - 删除未使用函数/变量
ODR 违规警告 - 重命名 ReorderMapData 避免冲突
Clang 特有警告 - 构造函数参数遮蔽修复
构建验证
✅ GCC Release 构建成功，无项目代码警告
⚠️ Clang Release 代码警告已修复，链接器配置问题需单独处理
备注
仅剩 quill 第三方库的 LTO 警告（库内部问题无法修复）
变更日志已保存至 changelog/2026-01-29-compiler-warnings-fix.md

+++ include/fqc/io/fastq_parser.h
@@ -131,7 +131,7 @@ struct ParserStats {
 
     /// @brief Get average read length.
     [[nodiscard]] double averageLength() const noexcept {
-        return totalRecords > 0 ? static_cast<double>(lengthSum) / totalRecords : 0.0;
+        return totalRecords > 0 ? static_cast<double>(lengthSum) / static_cast<double>(totalRecords) : 0.0;
     }
 
     /// @brief Reset statistics.
